feat(web): remove org scope from host detail core
