+++ apps/web/app/(dashboard)/hosts/[id]/compare/page.tsx
@@ -1,4 +1,3 @@
-import { getRequiredSession } from '@/lib/auth/session'
 import { CompareHostsClient } from './compare-client'
 
 interface Props {
@@ -7,10 +6,8 @@ interface Props {
 }
 
 export default async function CompareHostsPage({ params, searchParams }: Props) {
-  const session = await getRequiredSession()
   const { id: hostIdA } = await params
   const { with: hostIdB = '' } = await searchParams
-  const orgId = session.user.organisationId!
 
-  return <CompareHostsClient orgId={orgId} hostIdA={hostIdA} hostIdB={hostIdB} />
+  return <CompareHostsClient hostIdA={hostIdA} hostIdB={hostIdB} />
 }
