+++ apps/web/app/(dashboard)/hosts/[id]/page.tsx
@@ -2,6 +2,7 @@ import type { Metadata } from 'next'
 import { notFound } from 'next/navigation'
 import { getRequiredSession } from '@/lib/auth/session'
 import { getHost } from '@/lib/actions/agents'
+import { resolveCurrentActionScope } from '@/lib/actions/action-scope'
 import { REQUIRED_AGENT_VERSION } from '@/lib/agent/version'
 import { HostDetailClient } from './host-detail-client'
 
@@ -16,15 +17,15 @@ interface Props {
 export default async function HostDetailPage({ params }: Props) {
   const { id } = await params
   const session = await getRequiredSession()
-  const orgId = session.user.organisationId!
+  const scopeId = resolveCurrentActionScope(session)
 
-  const host = await getHost(orgId, id)
+  const host = await getHost(id)
   if (!host) notFound()
 
   return (
     <HostDetailClient
       host={host}
-      orgId={orgId}
+      scopeId={scopeId}
       currentUserId={session.user.id}
       userRole={session.user.role}
       latestAgentVersion={REQUIRED_AGENT_VERSION}
