+++ apps/web/app/(dashboard)/hosts/[id]/users/[accountId]/page.tsx
@@ -1,6 +1,5 @@
 import type { Metadata } from 'next'
 import { notFound } from 'next/navigation'
-import { getRequiredSession } from '@/lib/auth/session'
 import { getServiceAccount } from '@/lib/actions/service-accounts'
 import { LocalUserDetailClient } from './local-user-detail-client'
 
@@ -13,16 +12,13 @@ export default async function LocalUserDetailPage({
 }: {
   params: Promise<{ id: string; accountId: string }>
 }) {
-  const session = await getRequiredSession()
-  const orgId = session.user.organisationId!
   const { id: hostId, accountId } = await params
 
-  const result = await getServiceAccount(orgId, accountId, hostId)
+  const result = await getServiceAccount(accountId, hostId)
   if (!result) notFound()
 
   return (
     <LocalUserDetailClient
-      orgId={orgId}
       hostId={hostId}
       account={result.account}
       keys={result.keys}
